Fix some addresses being wrongly removed on DNS change
Previously, when multiple domains pointed to the same IP address and one of the domains
changed, the IP address was removed from the set, even if another domain still pointed
to the address.

Now, after a DNS update has run, we calculate the total difference per set to determine
which addresses to add or remove, touching only the elements that have changed.
Another improvement is that this way, the addresses for a domain are not cleared completely
on update, and the sets are modified only once per update, instead of once for every
domain that has changed.